Introduce the possibility to define different VCS backends
